Feat: Upload bike registration images in the background (#119)
# Description

Bike image uploads can take more than a few seconds with slow
connections, so this PR modifies the upload logic to always try and
upload in the background. That way if iOS terminates the app in the
background, the app will re-launch to handle a successful upload (by
saving the image to Swift Data so the bike's thumbnail is updated on the
main screen).

- Created separate postInBackground method on Client to perform post
requests in background.
- Added separate backgroundSession to Client with configuration set to
.background.
- Created BackgroundSessionDelegate that handles successful or failed
uploads.
- Created AppDelegate to store and later call the system's completion
handler (more info here:
https://developer.apple.com/documentation/foundation/urlsessiondelegate/urlsessiondidfinishevents(forbackgroundurlsession:).

## This can be tested in an iOS simulator by following these steps:
1. Go into the Settings app in the simulator, tap Developer, then toggle
on Fast App Termination under State Restoration Testing. This will
simulate app termination immediately after the app goes into the
background.
2. Run the following command in a terminal to view log messages after
the app is terminated: `xcrun simctl spawn booted log stream --level
debug --predicate 'subsystem == "org.bikeindex.ios"'`.
3. As soon as the alert message appears after registering a bike with
image in the simulator, go to the Home Screen and the app should then
automatically terminate itself. After some seconds, when the image
finishes uploading (or fails), it will run the app's session delegate
code in the background, and you should see log messages in the terminal.
4. Reopen the app, and if it was successful, the registered bike will
have an image thumbnail on the main screen. Tapping on the bike should
also show the uploaded image on the bike details view.